+++ Examples/TodoList/Classes/TodoListAppDelegate.m
@@ -22,12 +22,6 @@
 }
 
 
-- (void)dealloc
-{
-    [viewController release];
-    [window release];
-    [super dealloc];
-}
 
 
 @end