+++ HRCoder/HRCoder.m
@@ -1,7 +1,7 @@
 //
 //  HRCoder.m
 //
-//  Version 1.2.3
+//  Version 1.3
 //
 //  Created by Nick Lockwood on 24/04/2012.
 //  Copyright (c) 2011 Charcoal Design
@@ -33,6 +33,28 @@
 #import "HRCoder.h"
 
 
+#pragma GCC diagnostic ignored "-Wdirect-ivar-access"
+
+
+#import <Availability.h>
+#if !__has_feature(objc_arc)
+#error This class requires automatic reference counting
+#endif
+
+
+NSString *const HRCoderException = @"HRCoderException";
+NSString *const HRCoderClassNameKey = @"$class";
+NSString *const HRCoderRootObjectKey = @"$root";
+NSString *const HRCoderObjectAliasKey = @"$alias";
+
+
+@interface HRCoderAliasPlaceholder : NSObject
+
++ (HRCoderAliasPlaceholder *)placeholder;
+
+@end
+
+
 @interface NSObject (HRCoding_Private)
 
 - (id)unarchiveObjectWithHRCoder:(HRCoder *)coder;
@@ -55,7 +77,7 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"<%@>", NSStringFromClass([self class])];
+    return [NSString stringWithFormat:@"<%@>", [self class]];
 }
 
 @end
@@ -69,17 +91,10 @@
 @property (nonatomic, strong) NSString *keyPath;
 @property (nonatomic, strong) NSMutableData *data;
 
-+ (NSString *)classNameKey;
-
 @end
 
 
 @implementation HRCoder
-
-@synthesize stack = _stack;
-@synthesize knownObjects = _knownObjects;
-@synthesize unresolvedAliases = _unresolvedAliases;
-@synthesize keyPath = _keyPath;
 
 + (NSString *)classNameKey
 {
@@ -94,81 +109,66 @@
         self.stack = [NSMutableArray arrayWithObject:[NSMutableDictionary dictionary]];
         _knownObjects = [[NSMutableDictionary alloc] init];
         _unresolvedAliases = [[NSMutableDictionary alloc] init];
-        _outputFormat = NSPropertyListXMLFormat_v1_0;
+        _outputFormat = HRCoderFormatXML;
     }
     return self;
 }
 
-#if !__has_feature(objc_arc)
-
-- (void)dealloc
-{
-    [_stack release];
-    [_knownObjects release];
-    [_unresolvedAliases release];
-    [_keyPath release];
-    [_data release];
-    [super dealloc];
-}
-
-#endif
-
-
 #pragma mark -
 #pragma mark Unarchiving
 
-+ (id)unarchiveObjectWithPlist:(id)plist
++ (id)unarchiveObjectWithPlistOrJSON:(__unsafe_unretained id)plistOrJSON
 {
-    HRCoder *coder = [[self alloc] init];
-    
-#if !__has_feature(objc_arc)
-    [coder autorelease];
-#endif
-    
-    return [coder unarchiveRootObjectWithPlist:plist];
+    return [[[self alloc] init] unarchiveRootObjectWithPlistOrJSON:plistOrJSON];
 }
 
-+ (id)unarchiveObjectWithData:(NSData *)data
++ (id)unarchiveObjectWithData:(__unsafe_unretained NSData *)data
 {
-    //attempt to deserialise data as a plist
-    id plist = nil;
     if (data)
     {
-        NSPropertyListFormat format;
-        NSPropertyListReadOptions options = NSPropertyListMutableContainersAndLeaves;
-        plist = [NSPropertyListSerialization propertyListWithData:data options:options format:&format error:NULL];
+        //attempt to deserialize as plist
+        id plistOrJSON = [NSPropertyListSerialization propertyListWithData:data
+                                                                   options:NSPropertyListMutableContainers
+                                                                    format:NULL
+                                                                     error:NULL];
+        if (!plistOrJSON)
+        {
+            //attempt to deserialize as JSON
+            plistOrJSON = [NSJSONSerialization JSONObjectWithData:data options:NSJSONReadingMutableContainers error:NULL];
+        }
+        
+        //unarchive
+        return [self unarchiveObjectWithPlistOrJSON:plistOrJSON];
     }
-    
-    //unarchive
-    return [self unarchiveObjectWithPlist:plist];
+    return nil;
 }
 
-+ (id)unarchiveObjectWithFile:(NSString *)path
++ (id)unarchiveObjectWithFile:(__unsafe_unretained NSString *)path
 {
     //load the file
     return [self unarchiveObjectWithData:[NSData dataWithContentsOfFile:path]];
 }
 
-- (id)unarchiveRootObjectWithPlist:(id)plist
+- (id)unarchiveRootObjectWithPlistOrJSON:(__unsafe_unretained id)plistOrJSON
 {
     [_stack removeAllObjects];
     [_knownObjects removeAllObjects];
     [_unresolvedAliases removeAllObjects];
-    id rootObject = [plist unarchiveObjectWithHRCoder:self];
+    __autoreleasing id rootObject = [plistOrJSON unarchiveObjectWithHRCoder:self];
     if (rootObject)
     {
         _knownObjects[HRCoderRootObjectKey] = rootObject;
-        for (NSString *keyPath in _unresolvedAliases)
+        for (__unsafe_unretained NSString *keyPath in _unresolvedAliases)
         {
-            id aliasKeyPath = _unresolvedAliases[keyPath];
-            id aliasedObject = _knownObjects[aliasKeyPath];
-            id node = rootObject;
-            for (NSString *key in [keyPath componentsSeparatedByString:@"."])
+            __autoreleasing id aliasKeyPath = _unresolvedAliases[keyPath];
+            __autoreleasing id aliasedObject = _knownObjects[aliasKeyPath];
+            __autoreleasing id node = rootObject;
+            for (__unsafe_unretained NSString *key in [keyPath componentsSeparatedByString:@"."])
             {
-                id _node = nil;
+                __autoreleasing id _node = nil;
                 if ([node isKindOfClass:[NSArray class]])
                 {
-                    NSInteger index = [key integerValue];
+                    NSUInteger index = (NSUInteger)[key integerValue];
                     _node = node[index];
                     if (_node == [HRCoderAliasPlaceholder placeholder])
                     {
@@ -193,27 +193,57 @@
     return rootObject;
 }
 
-- (id)initForReadingWithData:(NSData *)data
+- (id)initForReadingWithData:(__unsafe_unretained NSData *)data
 {
     if ((self = [self init]))
     {
         //attempt to deserialise data as a plist
         if (data)
         {
-            //load plist
-            NSPropertyListReadOptions options = NSPropertyListMutableContainersAndLeaves;
-            id plist = [NSPropertyListSerialization propertyListWithData:data
-                                                                 options:options
-                                                                  format:&_outputFormat
-                                                                   error:NULL];
-            //only works if root object is a dictionary
-            if ([plist isKindOfClass:[NSDictionary class]])
+            //attempt to deserialize as plist
+            NSPropertyListFormat format;
+            NSError *error;
+            __autoreleasing id plistOrJSON = [NSPropertyListSerialization propertyListWithData:data
+                                                                                       options:NSPropertyListMutableContainers
+                                                                                        format:&format
+                                                                                         error:&error];
+            if (!plistOrJSON && error)
             {
-                [_stack addObject:plist];
+                //attempt to deserialize as JSON
+                plistOrJSON = [NSJSONSerialization JSONObjectWithData:data options:NSJSONReadingMutableContainers error:&error];
+                if (!plistOrJSON && error)
+                {
+                    [NSException raise:HRCoderException format:@"initForReadingWithData: method was unable to parse the supplied data."];
+                }
+                _outputFormat = HRCoderFormatJSON;
             }
             else
             {
-                [NSException raise:NSGenericException format:@"initForReadingWithData: method requires that the root object in the plist data is an NSDictionary. Decoding an %@ is not supported with this method. Try using unarchiveObjectWithData: instead.", [plist class]];
+                //convert to HRCoderFormat
+                switch (format)
+                {
+                    case NSPropertyListXMLFormat_v1_0:
+                    case NSPropertyListOpenStepFormat:
+                    {
+                        _outputFormat = HRCoderFormatXML;
+                        break;
+                    }
+                    case NSPropertyListBinaryFormat_v1_0:
+                    {
+                        _outputFormat = HRCoderFormatBinary;
+                        break;
+                    }
+                }
+            }
+            
+            //only works if root object is a dictionary
+            if ([plistOrJSON isKindOfClass:[NSDictionary class]])
+            {
+                [_stack addObject:plistOrJSON];
+            }
+            else
+            {
+                [NSException raise:HRCoderException format:@"initForReadingWithData: method requires that the root object in the plist data is an NSDictionary. Decoding an %@ is not supported with this method. Try using unarchiveObjectWithData: instead.", [plistOrJSON class]];
             }
         }
     }
@@ -231,37 +261,33 @@
 #pragma mark -
 #pragma mark Archiving
 
-+ (id)archivedPlistWithRootObject:(id)rootObject
++ (id)archivedPlistWithRootObject:(__unsafe_unretained id)rootObject
 {
     HRCoder *coder = [[self alloc] init];
-    id plist = [coder archiveRootObject:rootObject];
-    
-#if !__has_feature(objc_arc)
-    [coder release];
-#endif
-    
-    return plist;
+    return [coder archiveRootObject:rootObject];
 }
 
-+ (NSData *)archivedDataWithRootObject:(id)rootObject
++ (id)archivedJSONWithRootObject:(__unsafe_unretained id)rootObject
 {
-    NSMutableData *data = [NSMutableData data];
-    HRCoder *coder = [[self alloc] initForWritingWithMutableData:data];
+    HRCoder *coder = [[self alloc] init];
+    coder.outputFormat = HRCoderFormatJSON;
+    return [coder archiveRootObject:rootObject];
+}
+
++ (NSData *)archivedDataWithRootObject:(__unsafe_unretained id)rootObject
+{
+    __autoreleasing NSMutableData *data = [NSMutableData data];
+    __autoreleasing HRCoder *coder = [[self alloc] initForWritingWithMutableData:data];
     [coder archiveRootObject:rootObject];
-    
-#if !__has_feature(objc_arc)
-    [coder release];
-#endif
-    
     return data;
 }
 
-+ (BOOL)archiveRootObject:(id)rootObject toFile:(NSString *)path
++ (BOOL)archiveRootObject:(__unsafe_unretained id)rootObject toFile:(__unsafe_unretained NSString *)path
 {
     return [[self archivedDataWithRootObject:rootObject] writeToFile:path atomically:YES];
 }
 
-- (id)initForWritingWithMutableData:(NSMutableData *)data
+- (id)initForWritingWithMutableData:(__unsafe_unretained NSMutableData *)data
 {
     if ((self = [self init]))
     {
@@ -271,24 +297,36 @@
     return self;
 }
 
-- (id)archiveRootObject:(id)rootObject
+- (id)archiveRootObject:(__unsafe_unretained id)rootObject
 {
     [_knownObjects removeAllObjects];
     [self encodeRootObject:rootObject];
-    id plist = [_stack lastObject];
+    __autoreleasing id plistOrJSON = [_stack lastObject];
     [self finishEncoding];
-    return plist;
+    return plistOrJSON;
 }
 
 - (void)finishEncoding
 {
     if (_data)
     {
-        NSData *data = [NSPropertyListSerialization dataWithPropertyList:[_stack lastObject]
-                                                                  format:_outputFormat
-                                                                 options:0
-                                                                   error:NULL];
-        [_data setData:data];
+        switch (_outputFormat)
+        {
+            case HRCoderFormatXML:
+            case HRCoderFormatBinary:
+            {
+                [_data setData:[NSPropertyListSerialization dataWithPropertyList:[_stack lastObject]
+                                                                          format:(_outputFormat == HRCoderFormatXML)? NSPropertyListXMLFormat_v1_0: NSPropertyListBinaryFormat_v1_0
+                                                                         options:0
+                                                                           error:NULL]];
+                break;
+            }
+            case HRCoderFormatJSON:
+            {
+                [_data setData:[NSJSONSerialization dataWithJSONObject:[_stack lastObject] options:(NSJSONWritingOptions)0 error:NULL]];
+                break;
+            }
+        }
         self.data = nil;
     }
     [_knownObjects removeAllObjects];
@@ -304,20 +342,20 @@
     return YES;
 }
 
-- (BOOL)containsValueForKey:(NSString *)key
+- (BOOL)containsValueForKey:(__unsafe_unretained NSString *)key
 {
     return [_stack lastObject][key] != nil;
 }
 
-- (id)encodedObject:(id)object forKey:(NSString *)key
+- (id)encodedObject:(__unsafe_unretained id)object forKey:(__unsafe_unretained NSString *)key
 {
     if (object && key)
     {
         if (![object isKindOfClass:[NSNumber class]] &&
             ![object isKindOfClass:[NSDate class]] &&
-            (![object isKindOfClass:[NSString class]] || [object length] < 32))
+            (![object isKindOfClass:[NSString class]] || [(NSString *)object length] < 32))
         {
-            for (NSString *aliasKeyPath in _knownObjects)
+            for (__unsafe_unretained NSString *aliasKeyPath in _knownObjects)
             {
                 if (_knownObjects[aliasKeyPath] == object)
                 {
@@ -328,23 +366,23 @@
         }
         
         //encode object
-        NSString *oldKeyPath = _keyPath;
+        __autoreleasing NSString *oldKeyPath = _keyPath;
         self.keyPath = _keyPath? [_keyPath stringByAppendingPathExtension:key]: key;
         _knownObjects[_keyPath] = object;
-        id encodedObject = [object archivedObjectWithHRCoder:self];
+        __autoreleasing id encodedObject = [object archivedObjectWithHRCoder:self];
         self.keyPath = oldKeyPath;
         return encodedObject;
     }
     return nil;
 }
 
-- (void)encodeObject:(id)objv forKey:(NSString *)key
+- (void)encodeObject:(__unsafe_unretained id)objv forKey:(__unsafe_unretained NSString *)key
 {
-    id object = [self encodedObject:objv forKey:key];
+    __autoreleasing id object = [self encodedObject:objv forKey:key];
     if (object) [_stack lastObject][key] = object;
 }
 
-- (void)encodeRootObject:(id)rootObject
+- (void)encodeRootObject:(__unsafe_unretained id)rootObject
 {
     if (rootObject)
     {
@@ -353,9 +391,9 @@
     }
 }
 
-- (void)encodeConditionalObject:(id)objv forKey:(NSString *)key
+- (void)encodeConditionalObject:(id)objv forKey:(__unsafe_unretained NSString *)key
 {
-    for (id object in _knownObjects)
+    for (__unsafe_unretained id object in _knownObjects)
     {
         if (object == objv)
         {
@@ -365,56 +403,56 @@
     }
 }
 
-- (void)encodeBool:(BOOL)boolv forKey:(NSString *)key
+- (void)encodeBool:(BOOL)boolv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = @(boolv);
 }
 
-- (void)encodeInt:(int)intv forKey:(NSString *)key
+- (void)encodeInt:(int)intv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = @(intv);
 }
 
-- (void)encodeInt32:(int32_t)intv forKey:(NSString *)key
-{
-    [_stack lastObject][key] = [NSNumber numberWithLong:intv];
-}
-
-- (void)encodeInt64:(int64_t)intv forKey:(NSString *)key
+- (void)encodeInt32:(int32_t)intv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = @(intv);
 }
 
-- (void)encodeFloat:(float)realv forKey:(NSString *)key
+- (void)encodeInt64:(int64_t)intv forKey:(__unsafe_unretained NSString *)key
+{
+    [_stack lastObject][key] = @(intv);
+}
+
+- (void)encodeFloat:(float)realv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = @(realv);
 }
 
-- (void)encodeDouble:(double)realv forKey:(NSString *)key
+- (void)encodeDouble:(double)realv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = @(realv);
 }
 
-- (void)encodeBytes:(const uint8_t *)bytesp length:(NSUInteger)lenv forKey:(NSString *)key
+- (void)encodeBytes:(const uint8_t *)bytesp length:(NSUInteger)lenv forKey:(__unsafe_unretained NSString *)key
 {
     [_stack lastObject][key] = [NSData dataWithBytes:bytesp length:lenv];
 }
 
-- (id)decodeObject:(id)object forKey:(NSString *)key
+- (id)decodeObject:(id)object forKey:(__unsafe_unretained NSString *)key
 {
     if (object && key)
     {
         //new keypath
-        NSString *newKeyPath = _keyPath? [_keyPath stringByAppendingPathExtension:key]: key;
+        __autoreleasing NSString *newKeyPath = _keyPath? [_keyPath stringByAppendingPathExtension:key]: key;
         
         //check if object is an alias
         if ([object isKindOfClass:[NSDictionary class]])
         {
-            NSString *aliasKeyPath = ((NSDictionary *)object)[HRCoderObjectAliasKey];
+            __autoreleasing NSString *aliasKeyPath = ((NSDictionary *)object)[HRCoderObjectAliasKey];
             if (aliasKeyPath)
             {
                 //object alias
-                id decodedObject = _knownObjects[aliasKeyPath];
+                __autoreleasing id decodedObject = _knownObjects[aliasKeyPath];
                 if (!decodedObject)
                 {
                     _unresolvedAliases[newKeyPath] = aliasKeyPath;
@@ -425,9 +463,9 @@
         }
         
         //new object
-        NSString *oldKeyPath = _keyPath;
+        __autoreleasing NSString *oldKeyPath = _keyPath;
         self.keyPath = newKeyPath;
-        id decodedObject = [object unarchiveObjectWithHRCoder:self];
+        __autoreleasing id decodedObject = [object unarchiveObjectWithHRCoder:self];
         if (decodedObject) _knownObjects[_keyPath] = decodedObject;
         self.keyPath = oldKeyPath;
         return decodedObject;
@@ -435,44 +473,44 @@
     return nil;
 }
 
-- (id)decodeObjectForKey:(NSString *)key
+- (id)decodeObjectForKey:(__unsafe_unretained NSString *)key
 {
     return [self decodeObject:[_stack lastObject][key] forKey:key];
 }
 
-- (BOOL)decodeBoolForKey:(NSString *)key
+- (BOOL)decodeBoolForKey:(__unsafe_unretained NSString *)key
 {
     return [[_stack lastObject][key] boolValue];
 }
 
-- (int)decodeIntForKey:(NSString *)key
+- (int)decodeIntForKey:(__unsafe_unretained NSString *)key
 {
     return [[_stack lastObject][key] intValue];
 }
 
-- (int32_t)decodeInt32ForKey:(NSString *)key
+- (int32_t)decodeInt32ForKey:(__unsafe_unretained NSString *)key
 {
     return (int32_t)[[_stack lastObject][key] longValue];
 }
 
-- (int64_t)decodeInt64ForKey:(NSString *)key
+- (int64_t)decodeInt64ForKey:(__unsafe_unretained NSString *)key
 {
     return [[_stack lastObject][key] longLongValue];
 }
 
-- (float)decodeFloatForKey:(NSString *)key
+- (float)decodeFloatForKey:(__unsafe_unretained NSString *)key
 {
     return [[_stack lastObject][key] floatValue];
 }
 
-- (double)decodeDoubleForKey:(NSString *)key
+- (double)decodeDoubleForKey:(__unsafe_unretained NSString *)key
 {
     return [[_stack lastObject][key] doubleValue];
 }
 
-- (const uint8_t *)decodeBytesForKey:(NSString *)key returnedLength:(NSUInteger *)lengthp
+- (const uint8_t *)decodeBytesForKey:(__unsafe_unretained NSString *)key returnedLength:(NSUInteger *)lengthp
 {
-    NSData *data = [_stack lastObject][key];
+    __autoreleasing NSData *data = [_stack lastObject][key];
     *lengthp = [data length];
     return data.bytes;
 }
@@ -484,13 +522,12 @@
 
 - (id)unarchiveObjectWithHRCoder:(__unused HRCoder *)coder
 {
-    [NSException raise:NSGenericException format:@"%@ is not a supported HRCoder archive type", [self class]];
-    return nil;
+    return self;
 }
 
-- (id)archivedObjectWithHRCoder:(HRCoder *)coder
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
-    NSMutableDictionary *result = [NSMutableDictionary dictionary];
+    __autoreleasing NSMutableDictionary *result = [NSMutableDictionary dictionary];
     [coder.stack addObject:result];
     result[HRCoderClassNameKey] = NSStringFromClass([self classForCoder]);
     [(id <NSCoding>)[self replacementObjectForCoder:coder] encodeWithCoder:coder];
@@ -503,31 +540,26 @@
 
 @implementation NSDictionary(HRCoding)
 
-- (id)unarchiveObjectWithHRCoder:(HRCoder *)coder
+- (id)unarchiveObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
-    NSString *className = self[HRCoderClassNameKey];
+    __autoreleasing NSString *className = self[HRCoderClassNameKey];
     if (className)
     {
         //encoded object
         [coder.stack addObject:self];
-        Class class = NSClassFromString(className);
-        id object = [[[class alloc] initWithCoder:coder] awakeAfterUsingCoder:coder];
-        
-#if !__has_feature(objc_arc)
-        [object autorelease];
-#endif
-        
+        Class objectClass = NSClassFromString(className);
+        __autoreleasing id object = [[[objectClass alloc] initWithCoder:coder] awakeAfterUsingCoder:coder];
         [coder.stack removeLastObject];
         return object;
     }
     else
     {
         //ordinary dictionary
-        NSMutableDictionary *result = [NSMutableDictionary dictionary];
+        __autoreleasing NSMutableDictionary *result = [NSMutableDictionary dictionary];
 		[coder.stack addObject:self];
-        for (NSString *key in self)
+        for (__unsafe_unretained NSString *key in self)
         {
-            id object = [coder decodeObjectForKey:key];
+            __autoreleasing id object = [coder decodeObjectForKey:key];
             if (object) result[key] = object;
         }
 		[coder.stack removeLastObject];
@@ -535,14 +567,13 @@
     }
 }
 
-- (id)archivedObjectWithHRCoder:(HRCoder *)coder
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
-    NSMutableDictionary *result = [NSMutableDictionary dictionary];
+    __autoreleasing NSMutableDictionary *result = [NSMutableDictionary dictionary];
     [coder.stack addObject:result];
-    for (NSString *key in self)
-    {
-        [coder encodeObject:self[key] forKey:key];
-    }
+    [self enumerateKeysAndObjectsUsingBlock:^(__unsafe_unretained id key, __unsafe_unretained id obj, __unused BOOL *stop) {
+        [coder encodeObject:obj forKey:key];
+    }];
     [coder.stack removeLastObject];
     return result;
 }
@@ -552,9 +583,9 @@
 
 @implementation NSArray(HRCoding)
 
-- (id)unarchiveObjectWithHRCoder:(HRCoder *)coder
+- (id)unarchiveObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
-    NSMutableArray *result = [NSMutableArray array];
+    __autoreleasing NSMutableArray *result = [NSMutableArray array];
     for (NSUInteger i = 0; i < [self count]; i++)
     {
         NSString *key = [@(i) description];
@@ -565,7 +596,7 @@
     return result;
 }
 
-- (id)archivedObjectWithHRCoder:(HRCoder *)coder
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
     NSMutableArray *result = [NSMutableArray array];
     for (NSUInteger i = 0; i < [self count]; i++)
@@ -587,23 +618,12 @@
     return self;
 }
 
-- (id)archivedObjectWithHRCoder:(__unused HRCoder *)coder
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
-    return self;
-}
-
-@end
-
-
-@implementation NSData(HRCoding)
-
-- (id)unarchiveObjectWithHRCoder:(__unused HRCoder *)coder
-{
-    return self;
-}
-
-- (id)archivedObjectWithHRCoder:(__unused HRCoder *)coder
-{
+    if ([self classForCoder] == [NSMutableString class])
+    {
+        return [super archivedObjectWithHRCoder:coder];
+    }
     return self;
 }
 
@@ -625,6 +645,25 @@
 @end
 
 
+@implementation NSData(HRCoding)
+
+- (id)unarchiveObjectWithHRCoder:(__unused HRCoder *)coder
+{
+    return self;
+}
+
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
+{
+    if (coder.outputFormat == HRCoderFormatJSON || [self classForCoder] == [NSMutableData class])
+    {
+        return [super archivedObjectWithHRCoder:coder];
+    }
+    return self;
+}
+
+@end
+
+
 @implementation NSDate(HRCoding)
 
 - (id)unarchiveObjectWithHRCoder:(__unused HRCoder *)coder
@@ -632,8 +671,31 @@
     return self;
 }
 
-- (id)archivedObjectWithHRCoder:(__unused HRCoder *)coder
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
 {
+    if (coder.outputFormat == HRCoderFormatJSON)
+    {
+        return [super archivedObjectWithHRCoder:coder];
+    }
+    return self;
+}
+
+@end
+
+
+@implementation NSNull(HRCoding)
+
+- (id)unarchiveObjectWithHRCoder:(__unused HRCoder *)coder
+{
+    return self;
+}
+
+- (id)archivedObjectWithHRCoder:(__unsafe_unretained HRCoder *)coder
+{
+    if (coder.outputFormat != HRCoderFormatJSON)
+    {
+        return [super archivedObjectWithHRCoder:coder];
+    }
     return self;
 }